refactor(admin): 케밥케이스로 이름 수정 #87
